pb can role user

+++ app/Http/Controllers/UserController.php
@@ -91,6 +91,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize("delete", $user);
         $user->delete();
         return redirect()->route("user.index");
     }
